port most backend functionality, further cleanups, basic electron support
