+++ helpers/websockets.js
@@ -8,7 +8,7 @@ const config = require('config');
 
 const WebSocketServer = require('ws').Server;
 
-const RedisConnection = require('ioredis');
+//const RedisConnection = require('ioredis');
 const async = require('async');
 const _ = require("underscore");
 const crypto = require('crypto');
