+++ package.json
@@ -3,8 +3,7 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "start": "nodemon -e .js,.html bin/www",
-    "test": "mocha"
+    "start": "electron ."
   },
   "engines": {
     "node": ">=7.8.0"
@@ -12,48 +11,26 @@
   "dependencies": {
     "archiver": "1.3.0",
     "async": "2.3.0",
-    "aws-sdk": "2.39.0",
     "basic-auth": "1.1.0",
     "bcryptjs": "2.4.3",
     "body-parser": "~1.17.1",
     "cheerio": "0.22.0",
     "config": "1.25.1",
     "cookie-parser": "~1.4.3",
-    "csurf": "1.9.0",
-    "debug": "~2.6.3",
     "electron": "^1.8.4",
     "execSync": "latest",
     "express": "~4.13.0",
-    "extract-zip": "^1.6.6",
     "file-type": "^7.6.0",
     "glob": "7.1.1",
     "gm": "1.23.0",
-    "googleapis": "18.0.0",
-    "gulp": "^3.9.1",
-    "gulp-concat": "2.6.0",
-    "gulp-express": "0.3.0",
-    "gulp-nodemon": "*",
-    "gulp-sass": "^2.0.3",
-    "gulp-uglify": "^1.5.1",
-    "gulp-util": "^3.0.6",
     "helmet": "^3.5.0",
     "i18n-2": "0.6.3",
-    "ioredis": "2.5.0",
-    "lodash": "^4.3.0",
     "log-timestamp": "latest",
-    "md5": "2.2.1",
-    "mime-types": "^2.1.18",
+    "morgan": "1.8.1",
     "mock-aws-s3": "^2.6.0",
     "moment": "^2.19.3",
-    "mongoose": "4.9.3",
-    "morgan": "1.8.1",
     "node-phantom-simple": "2.2.4",
-    "node-sass-middleware": "0.11.0",
-    "pdfkit": "0.8.0",
     "phantomjs-prebuilt": "2.1.14",
-    "pm2": "latest",
-    "qr-image": "3.2.0",
-    "raven": "1.2.0",
     "read-chunk": "^2.1.0",
     "request": "2.81.0",
     "sanitize-html": "^1.11.1",
@@ -66,29 +43,10 @@
     "underscore": "1.8.3",
     "uuid": "^3.2.1",
     "validator": "7.0.0",
-    "weak": "1.0.1",
     "ws": "2.2.3"
   },
-  "devDependencies": {
-    "express": "^4.13.3",
-    "gulp": "^3.9.1",
-    "gulp-clean": "^0.3.2",
-    "gulp-concat": "^2.6.0",
-    "gulp-express": "^0.3.0",
-    "gulp-fingerprint": "^0.3.2",
-    "gulp-nodemon": "^2.0.4",
-    "gulp-rev": "^7.1.2",
-    "gulp-rev-all": "^0.9.7",
-    "gulp-rev-replace": "^0.4.3",
-    "gulp-sass": "^3.1.0",
-    "gulp-uglify": "^2.1.2",
-    "nodemon": "1.11.0",
-    "should": "^11.2.1",
-    "supertest": "^3.0.0",
-    "winston": "^2.3.1"
-  },
+  "main": "app.js",
   "description": "",
-  "main": "Gulpfile.js",
   "directories": {},
   "repository": {
     "type": "git",
