+++ routes/api/space_digest.js
@@ -6,12 +6,10 @@ require('../../models/db');
 var async = require('async');
 var fs = require('fs');
 var _ = require("underscore");
-var mongoose = require("mongoose");
 var request = require('request');
 var url = require("url");
 var path = require("path");
 var crypto = require('crypto');
-var qr = require('qr-image');
 var glob = require('glob');
 var gm = require('gm');
